adding alert for logout and delete anggota

+++ lib/screen/anggota.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tugas_login/dataSource/api.dart';
 import 'package:get_storage/get_storage.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 final _storage = GetStorage();
 
@@ -9,6 +10,13 @@ class anggotaPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        leading: IconButton(
+          icon: Icon(Icons.arrow_back),
+          onPressed: () {
+            Navigator.pop(context);
+            Navigator.pushReplacementNamed(context, '/home');
+          },
+        ),
         title: const Center(
             child: Text("ANGGOTA",
                 style: TextStyle(
@@ -70,8 +78,35 @@ class anggotaPage extends StatelessWidget {
                               getEditAnggotaDetail(
                                   context, _storage.read('id_${index + 1}'));
                             } else if (value == 'Delete') {
-                              deleteUser(
-                                  context, _storage.read('id_${index + 1}'));
+                              showDialog(
+                                  context: context,
+                                  builder: (context) => AlertDialog(
+                                        actions: [
+                                          TextButton(
+                                            onPressed: () {
+                                              deleteUser(
+                                                  context,
+                                                  _storage
+                                                      .read('id_${index + 1}'));
+                                            },
+                                            child: const Text('ya'),
+                                          ),
+                                          TextButton(
+                                            onPressed: () {
+                                              Navigator.of(context).pop();
+                                            },
+                                            child: const Text('tidak'),
+                                          )
+                                        ],
+                                        contentPadding: EdgeInsets.all(20.0),
+                                        content: Text(
+                                          'Apakah anda yakin?',
+                                          style: GoogleFonts.poppins(
+                                              textStyle: TextStyle(
+                                                  color: Colors.black,
+                                                  fontSize: 20)),
+                                        ),
+                                      ));
                             } else if (value == 'Detail') {
                               showDialog(
                                 context: context,
@@ -84,23 +119,126 @@ class anggotaPage extends StatelessWidget {
                                         child: const Text('close'),
                                       ),
                                     ],
-                                    title: Text('Detail Anggota'),
-                                    contentPadding: EdgeInsets.all(20.0),
-                                    content: Column(
-                                      children: [
-                                        Text(
-                                            'Nama: ${_storage.read('nama_${index + 1}')}'),
-                                        Text(
-                                            'Nomor Induk: ${_storage.read('nomor_induk_${index + 1}')}'),
-                                        Text(
-                                            'Telepon: ${_storage.read('telepon_${index + 1}')}'),
-                                        Text(
-                                            'Status Aktif: ${_storage.read('status_aktif_${index + 1}')}'),
-                                        Text(
-                                            'Alamat: ${_storage.read('alamat_${index + 1}')}'),
-                                        Text(
-                                            'Tanggal Lahir: ${_storage.read('tgl_lahir_${index + 1}')}'),
-                                      ],
+                                    title:
+                                        Center(child: Text('Detail Anggota')),
+                                    contentPadding: EdgeInsets.only(
+                                        top: 20,
+                                        bottom: 50,
+                                        right: 20,
+                                        left: 20),
+                                    content: SingleChildScrollView(
+                                      child: Column(
+                                        children: [
+                                          Row(
+                                            mainAxisAlignment:
+                                                MainAxisAlignment.start,
+                                            children: [
+                                              Text('Nama:',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              '${_storage.read('nama_${index + 1}')}',
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                          Row(
+                                            children: [
+                                              Text('Nomor Induk: ',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              '${_storage.read('nomor_induk_${index + 1}')}',
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                          Row(
+                                            children: [
+                                              Text('Telepon: ',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              '${_storage.read('telepon_${index + 1}')}',
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                          Row(
+                                            children: [
+                                              Text('Status Aktif:',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              _storage.read(
+                                                          'status_aktif_${index + 1}') ==
+                                                      1
+                                                  ? "Aktif"
+                                                  : "Tidak Aktif",
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                          Row(
+                                            children: [
+                                              Text('Alamat:',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              '${_storage.read('alamat_${index + 1}')}',
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                          Row(
+                                            children: [
+                                              Text('Tanggal Lahir:',
+                                                  style: GoogleFonts.poppins(
+                                                      textStyle: TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 20))),
+                                            ],
+                                          ),
+                                          Text(
+                                              '${_storage.read('tgl_lahir_${index + 1}')}',
+                                              style: GoogleFonts.poppins(
+                                                  textStyle: TextStyle(
+                                                      color: Colors.black,
+                                                      fontSize: 20))),
+                                        ],
+                                      ),
                                     )),
                               );
                             }
